+++ installation-bundle/src/InstallTool.php
@@ -14,6 +14,7 @@ use Contao\Backend;
 use Contao\Config;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
@@ -36,22 +37,22 @@ class InstallTool
     private $rootDir;
 
     /**
-     * @var string
+     * @var LoggerInterface
      */
-    private $logDir;
+    private $logger;
 
     /**
      * Constructor.
      *
-     * @param Connection $connection
-     * @param string     $rootDir
-     * @param string     $logDir
+     * @param Connection      $connection
+     * @param string          $rootDir
+     * @param LoggerInterface $logger
      */
-    public function __construct(Connection $connection, $rootDir, $logDir)
+    public function __construct(Connection $connection, $rootDir, LoggerInterface $logger)
     {
         $this->connection = $connection;
         $this->rootDir = $rootDir;
-        $this->logDir = $logDir;
+        $this->logger = $logger;
     }
 
     /**
@@ -135,13 +136,15 @@ class InstallTool
      */
     public function canConnectToDatabase($name)
     {
-        if (null === $this->connection) {
+        if (null === $name || null === $this->connection) {
             return false;
         }
 
         try {
             $this->connection->connect();
         } catch (\Exception $e) {
+            $this->logException($e);
+
             return false;
         }
 
@@ -150,6 +153,8 @@ class InstallTool
         try {
             $this->connection->query('use '.$quotedName);
         } catch (DBALException $e) {
+            $this->logException($e);
+
             return false;
         }
 
@@ -394,16 +399,6 @@ class InstallTool
      */
     public function logException(\Exception $e)
     {
-        error_log(
-            sprintf(
-                "PHP Fatal error: %s in %s on line %s\n%s\n",
-                $e->getMessage(),
-                $e->getFile(),
-                $e->getLine(),
-                $e->getTraceAsString()
-            ),
-            3,
-            $this->logDir.'/prod-'.date('Y-m-d').'.log'
-        );
+        $this->logger->critical('An exception occurred.', ['exception' => $e]);
     }
 }