+++ installation-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao installation bundle change log
 
+### DEV
+
+ * Log database connection errors (see contao/core-bundle#1324).
+
 ### 4.4.12 (2018-01-03)
 
  * The assets:install command requires the application to be set in Symfony 3.4 (see #81).